fix admin screening detail page

+++ src/routes/admin/manage-halls/+page.svelte
@@ -10,6 +10,10 @@
 
 	// Saal löschen
 	async function deleteHall(hall_id: number) {
+		if (!confirm('Sind Sie sicher, dass Sie diesen Saal löschen möchten?')) {
+			return;
+		}
+
 		const response = await fetch(`/admin/manage-halls/${hall_id}`, {
 			method: 'DELETE'
 		});
